added fallback material and moved helper functions
